+++ src/inmanta/server/agentmanager.py
@@ -21,7 +21,7 @@ from tornado import gen
 from tornado import locks
 
 from inmanta.config import Config
-from inmanta import data, methods
+from inmanta import data
 from inmanta.server import protocol, SLICE_AGENT_MANAGER
 from inmanta.asyncutil import retry_limited
 from . import config as server_config
@@ -33,9 +33,9 @@ import time
 import sys
 import subprocess
 import uuid
-from inmanta.server.protocol import ServerSlice
+from inmanta.server.protocol import ServerSlice, SessionListener
 from inmanta.server import config as opt
-from inmanta.protocol import encode_token
+from inmanta.protocol import encode_token, methods
 from inmanta.resources import Id
 
 
@@ -79,7 +79,7 @@ set_parameters
 """
 
 
-class AgentManager(ServerSlice):
+class AgentManager(ServerSlice, SessionListener):
     '''
     This class contains all server functionality related to the management of agents
     '''
@@ -110,9 +110,9 @@ class AgentManager(ServerSlice):
     @gen.coroutine
     def prestart(self, server):
         yield ServerSlice.prestart(self, server)
-        self._server = server.get_endpoint("server")
+        self._server = server.get_slice("server")
         self._server_storage = self._server._server_storage
-        server.get_endpoint("session").add_listener(self)
+        server.get_slice("session").add_listener(self)
 
     def new_session(self, session):
         self.add_future(self.register_session(session, datetime.now()))
@@ -339,17 +339,17 @@ class AgentManager(ServerSlice):
                 errhandle.close()
 
     # External APIS
-    @protocol.handle(methods.NodeMethod.get_agent_process, agent_id="id")
+    @protocol.handle(methods.get_agent_process, agent_id="id")
     @gen.coroutine
     def get_agent_process(self, agent_id):
         return (yield self.get_agent_process_report(agent_id))
 
-    @protocol.handle(methods.ServerAgentApiMethod.trigger_agent, agent_id="id", env="tid")
+    @protocol.handle(methods.trigger_agent, agent_id="id", env="tid")
     @gen.coroutine
     def trigger_agent(self, env, agent_id):
         raise NotImplementedError()
 
-    @protocol.handle(methods.NodeMethod.list_agent_processes)
+    @protocol.handle(methods.list_agent_processes)
     @gen.coroutine
     def list_agent_processes(self, environment, expired):
         if environment is not None:
@@ -382,7 +382,7 @@ class AgentManager(ServerSlice):
 
         return 200, {"processes": processes}
 
-    @protocol.handle(methods.ServerAgentApiMethod.list_agents, env="tid")
+    @protocol.handle(methods.list_agents, env="tid")
     @gen.coroutine
     def list_agents(self, env):
         if env is not None:
@@ -393,9 +393,9 @@ class AgentManager(ServerSlice):
 
         return 200, {"agents": [a.to_dict() for a in ags], "servertime": datetime.now().isoformat()}
 
-    @protocol.handle(methods.AgentRecovery.get_state, env="tid")
+    @protocol.handle(methods.get_state, env="tid")
     @gen.coroutine
-    def get_state(self, env: uuid.UUID, sid: uuid.UUID, agent: str):
+    def get_state(self, env: data.Environment, sid: uuid.UUID, agent: str):
         tid = env.id
         if isinstance(tid, str):
             tid = uuid.UUID(tid)